+++ includes/wf_time.c
@@ -1,11 +1,12 @@
 #include "./time.h"
 #include <time.h>
 #include <sys/time.h>
+#include <stdint.h>
 
 
 /* Returns the current timestamp in microseconds */
-unsigned long long now_ns() {
+uint64_t now_ns() {
 	struct timeval v;
 	gettimeofday(&v,NULL);
-	return (unsigned long long)(v.tv_sec*1000000000 + v.tv_usec*1000); 
+	return (uint64_t)(v.tv_sec*1000000000 + v.tv_usec*1000); 
 }
